Use `spawn` instead of `exec` to start mplayer
When started with `exec` mplayer was stopping with an error about the
stdout buffer being exceded. Using `spawn` seems to fix this:
https://stackoverflow.com/a/26995407/1107844

I've also changed back to sending messages to mplayer over STDIN instead
of using a fifo. It probably is simpler this way... means we don't have
the fifo file that we need to clean up.